pyproject.toml
[project]
name = "gym-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "openai",
    "gspread>=6.2.1",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "uvicorn>=0.40.0",
]
